fix(*): move Modal controls into App
